Hit and Run

This function provides a wrapper for the har function of the hit-and-run package

hit_and_run(A, b, x0, points, thin = 1, burn = 0, chains = 1)

Arguments

A
is the lhs of Ax <= b

b
is the rhs of Ax <= b

x0
is the starting point (a list of points)

points
is the number of points we want to sample

thin
every thin-th point is stored

burn
the first burn points are deleted

chains
is the number of chains we run

Value

a list of chains of the sampled points, each chain being a matrix object with each column as a point

Examples

if (FALSE) {
## note that this Ax <= b is different from Ax=b that the 
## user specifies for walkr (see transformation section in vignette) 
hit_and_run(A = A, b = b, x0, points = 100, thin = 1, burn = 0, chains = 1)
}